fix: prevent redirect loop
